+++ TRRandomizerCore/Editors/TR2RemasteredEditor.cs
@@ -2,6 +2,7 @@
 using TRGE.Core;
 using TRLevelControl.Model;
 using TRRandomizerCore.Helpers;
+using TRRandomizerCore.Processors;
 using TRRandomizerCore.Randomizers;
 
 namespace TRRandomizerCore.Editors;
@@ -62,6 +63,12 @@ public class TR2RemasteredEditor : TR2ClassicEditor
 
         // Environment randomizer always runs
         target += numLevels * 2;
+
+        // Level sequencing checks
+        if (Settings.RandomizeGameMode)
+        {
+            target += numLevels;
+        }
 
         return target;
     }
@@ -221,6 +228,20 @@ public class TR2RemasteredEditor : TR2ClassicEditor
                 SaveMonitor = monitor,
                 Settings = Settings,
             }.Randomize(Settings.TextureSeed);
+        }
+
+        if (!monitor.IsCancelled && Settings.RandomizeGameMode)
+        {
+            monitor.FireSaveStateBeginning(TRSaveCategory.Custom, "Running level sequence checks");
+            new TR2RSequenceProcessor
+            {
+                ScriptEditor = scriptEditor,
+                Levels = levels,
+                BasePath = wipDirectory,
+                BackupPath = backupDirectory,
+                SaveMonitor = monitor,
+                Settings = Settings,
+            }.Run();
         }
 
         monitor.FireSaveStateBeginning(TRSaveCategory.Custom, "Finalizing tasks - please wait");