+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased](https://github.com/LostArtefacts/TR-Rando/compare/V1.9.1...master) - xxxx-xx-xx
+- added support for level sequence randomization in TR1R and TR2R (#756)
 - added options to use textures from specific game areas only in TRR texture randomization (#726)
 - changed vehicle randomization in TR2 so that it is now optional within item randomization (#750)
 - fixed key item softlocks in remastered New Game+ when using shuffled item mode (#732, #734)